requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
xlrd==2.0.1
openpyxl==3.1.2
SQLAlchemy==2.0.25
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
stripe==9.6.0
python-dotenv==1.0.1
email-validator==2.2.0
